requirements.txt
alpaca-py==0.8.1
appdirs==1.4.4
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
finnhub-python==2.4.17
frozendict==2.3.7
html5lib==1.1
idna==3.4
joblib==1.2.0
lxml==4.9.2
msgpack==1.0.5
multitasking==0.0.11
nltk==3.8.1
numpy==1.24.2
pandas==1.5.3
pycparser==2.21
pydantic==1.10.7
python-dateutil==2.8.2
pytz==2023.3
regex==2023.3.23
requests==2.28.2
six==1.16.0
soupsieve==2.4.1
sseclient-py==1.7.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
webencodings==0.5.1
websockets==10.4
yfinance==0.2.18
